Make Types and impelemented bets logic
