+++ blackjack.cabal
@@ -25,8 +25,10 @@ source-repository head
 
 library
   exposed-modules:
+      Bets
       Deck
       Game
+      Types
   other-modules:
       Paths_blackjack
   hs-source-dirs:
@@ -36,6 +38,7 @@ library
       base >=4.7 && <5
     , lens
     , random
+    , split
     , transformers
   default-language: Haskell2010
 
@@ -49,6 +52,7 @@ executable blackjack-exe
     , blackjack
     , lens
     , random
+    , split
     , transformers
   default-language: Haskell2010
 
@@ -65,5 +69,6 @@ test-suite blackjack-test
     , blackjack
     , lens
     , random
+    , split
     , transformers
   default-language: Haskell2010
