+++ package.yaml
@@ -23,6 +23,7 @@ dependencies:
 - base >= 4.7 && < 5
 - random
 - transformers
+- split
 - lens
 ghc-options:
 - -Wall
