+++ src/Game.hs
@@ -1,50 +1,11 @@
 module Game where
 
 import Deck 
+import Types
+import Bets
 import System.Random
 import Control.Monad.Trans.State
-
-data Game = Game
-    { deck        :: Deck
-    , discard     :: Deck
-    , dealer      :: Dealer
-    , players     :: [Player]
-    , penetration :: Int
-    , gen         :: StdGen}
-    deriving (Show)
-
-type Money = Int
-
-type Hand = [Card]
-
-data Player = Player { playerName :: String
-                     , hands      :: [Hand]
-                     , bankroll   :: Money
-                     , bet        :: [Money] }
-
-data Dealer = Dealer { dealerName :: String
-                     , hand       :: Hand
-                     , hiddenHand :: Hand }
-     deriving (Eq)
-
-
-
-instance Show Player where
-    show (Player name hs br bs) = "Player: " ++ name ++
-                                  "\nHands: " ++ show hs ++
-                                  "\nBets: " ++ show bs ++
-                                  "\nBankroll: " ++ show br
-                                  
-instance Show Dealer where
-    show (Dealer name hs hh)    = if null hh
-                                    then "Dealer: " ++ name ++
-                                         "\nHand: " ++ show hs
-                                    else "Dealer: " ++ name ++
-                                         "\nHand: " ++ hideCard
-        where
-            hideCard = takeWhile (/=',') (show (hs ++ hh)) ++ ", XXX]"
-
-type GameT m = StateT Game m
+import Control.Monad.IO.Class (liftIO)
 
 printTableValue :: [Card] -> IO ()
 printTableValue cs = putStrLn $ "The value of cards is " ++ show vals
@@ -87,11 +48,8 @@ handValue hs = if isSoft hs
 dealOpeningHands :: Monad m => GameT m ()
 dealOpeningHands = do
     game <- get
-    let players' = players game
-    let deck' = deck game
-    let dealer' = dealer game
-    let (newPlayers, newDeck) = dealFirstCard ([], deck') players'
-    let (dealer1, newDeck') = dealerCard newDeck dealer'
+    let (newPlayers, newDeck) = dealFirstCard ([], deck game) (players game)
+    let (dealer1, newDeck') = dealerCard newDeck (dealer game)
     let (startPlayers, newDeck'') = startingHands ([], newDeck') newPlayers
     let (dealer2, newDeck''') = dealerCardH newDeck'' dealer1
     let game' = game { dealer = dealer2, players = startPlayers, deck = newDeck''' }
@@ -160,3 +118,34 @@ drawNInitial n (c, dk) = drawNInitial (n-1) (c ++ [[c']], dk')
 drawCard :: Deck -> (Card, Deck)
 drawCard (c:d) = (c, d)
 drawCard [] = error "Cannot draw from an empty deck"
+
+playGame :: GameT IO ()
+playGame = do
+    game <- get
+    playerWBets <- liftIO $ getBets (players game)
+    let game' = game { players = playerWBets }
+    liftIO $ print game'
+    put game'
+    -- cleanup bets
+    -- deal opening hands
+    -- if Ace/Ten, ask for insurance bets
+        -- check for dealer 21
+        -- resolve bets if 21
+    -- for each player
+        -- process each hand in order
+        -- for each hand, if it's a blackjack, tell them and pay them the payout rate
+        -- if not, ask if they want to hit, stand, double (if they have enough money) or 
+            -- split (if the two cards have the same numerical value and they have enough money and the number of hands is less than 4)
+        -- if they split, take the second card and make a new hand after the previous, debt their account and add the bet
+            -- deal the current hand another card and repeat
+            -- a two-card 21 is not a blackjack and pays 1:1
+        -- if they double down, double their bet, debt the account and give them one card and stop prompts for the hand
+        -- if they hit, give them another card and repeat
+        -- if they stand, move to the next hand
+        -- if anytime the player's hand totals more than 21, immediate take the bet and move the cards to the discard pile
+    -- the dealer then moves the hidden card to their hand and continues to hit until the stopping criteria
+    -- all remaining bets are resolved (push = money goes back, win = double money, loss = clearing of bets)
+    -- move all cards to the discard pile
+    -- if the size of the deck is less than the penetration size, shuffle all cards
+        -- reset the cut card
+    -- if at anytime the deck runs out, shuffle the discards
